feat: clickable PostCard component, temporary styling
