+++ README.md
@@ -1,3 +1,15 @@
 # riv.systems
 
 This project serves as a portfolio/blog hybrid site.
+
+## GitHub CMS
+
+The cms directory serves as a repository of content to be rendered on the frontend.
+
+#### Terminology
+
+The structure of the CMS is comprised of the following levels: Section, Category, and Content. These levels can be seen similarly for intiatives.
+
+1. Section: Sections are considered the first layer, and can includes titles such as "porfolio", "journal", etc.
+2. Category: The second layer, and are categories for a given section. For example, categories that can go under portfolio include "UX/UI Design", "Web Development", "Graphic Design", etc.
+3. Content: The blog posts for a given category. They are writen in markdown and rednered as JSX.
