+++ components/PostCard/PostCard.module.css
@@ -1,0 +1,6 @@
+.container {
+  width: 100%;
+  max-width: 250px;
+  border: 1px white solid;
+  padding: 10px;
+}
